Fix Coverity errors
Fix pointers to outside local scope
Fix out-of-bounds memory accesses
Fix incorrect initialisation of blank string
Remove Coverity false positive
Fix Lithium incorrect logical OR same on both sides
Fix incorrect sizing of payload processor index
Fix incorrect iterator usage in PayloadSectionManager
Initialise uninitialised members of IR sensor and AntennaBurnerInfo
Remove logically dead code in RTC valid time check

+++ src/payload_processor/payload_processor.cpp
@@ -35,7 +35,7 @@
 PayloadProcessor::PayloadProcessor() : payload_section_manager() {}
 
 bool PayloadProcessor::ParseAndExecuteUplinks(byte* payload) {
-    byte current_index = 0;
+    uint16_t current_index = 0;
 
     while (current_index < Lithium::kMaxReceivedUplinkSize &&
            !(payload[current_index] == kEndTerminator &&
@@ -117,7 +117,7 @@ Uplink* PayloadProcessor::CreateUplink(uint16_t command_code, byte* payload) {
     }
 }
 
-bool PayloadProcessor::ParseNextUplinkAndExecute(uint8_t& index,
+bool PayloadProcessor::ParseNextUplinkAndExecute(uint16_t& index,
                                                  byte* payload) {
     uint16_t command_code = static_cast<uint16_t>(payload[index]) |
                             (static_cast<uint16_t>(payload[index + 1]) << 8);
